remove useless codes & rename file
