+++ src/components/Header/MenusComp/index.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import classNames from 'classnames'
 import dayjs from 'dayjs'
 import { Link } from '../../Link'
-import { MobileMenuItem } from './styled'
+import { MobileMenuItem } from './MobileMenuItem'
 import styles from './index.module.scss'
 import { LanguageModal } from '../LanguageComp/LanguageModal'
 import { CKBNodeModal } from '../CKBNodeComp/CKBNodeModal'
