+++ src/components/Text/HighLightPanel.tsx
