+++ src/components/TransactionItem/styles.module.scss
@@ -34,9 +34,6 @@
 .transactionPanel {
   width: 100%;
   margin-top: 4px;
-
-  // border-radius: ${(props: { circleCorner: CircleCorner }) =>
-  //   `${props.circleCorner.top ? '6px 6px' : '0 0'}${props.circleCorner.bottom ? ' 6px 6px' : ' 0 0'}`};
   box-shadow: 2px 2px 6px 0 #dfdfdf;
   background-color: #fff;
   padding: 10px 40px 15px;
