+++ src/pages/BlockList/index.tsx
@@ -6,7 +6,7 @@ import { Link } from '../../components/Link'
 import { parseSimpleDate } from '../../utils/date'
 import Content from '../../components/Content'
 import { TableContentItem, TableMinerContentItem } from '../../components/Table'
-import { TableTitleRow, TableContentRow, TableTitleRowItem } from '../../components/Table/styled'
+import { TableTitleRow, TableContentRow, TableTitleRowItem } from '../../components/Table/TableComp'
 import { deprecatedAddrToNewAddr, shannonToCkb } from '../../utils/util'
 import { DELAY_BLOCK_NUMBER } from '../../constants/common'
 import { localeNumberString } from '../../utils/number'
