+++ src/pages/Script/ScriptsComp.tsx
@@ -8,7 +8,7 @@ import { RefreshCw } from 'lucide-react'
 import Pagination from '../../components/Pagination'
 import TransactionItem from '../../components/TransactionItem/index'
 import { explorerService } from '../../services/ExplorerService'
-import { TransactionCellInfoPanel } from '../Transaction/TransactionCell/styled'
+import { TransactionCellInfoPanel } from '../Transaction/TransactionCell/TransactionCellComp'
 import SimpleButton from '../../components/SimpleButton'
 import SimpleModal from '../../components/Modal'
 import { localeNumberString } from '../../utils/number'
